Refactor exception handling
Exceptions lose ex-data when they are thrown from pod. To fix this we
now manually serialize exceptions in pod and rethrow them in Boot pod.
This serialization should keep stack-trace and causes.

In pod, Cljs exceptions are modified - that is new ExceptionInfo is
created with the same cause. In exception data and message the file path
is modified so that it point to the original source-path instead of to
temp-dir.